Added compress only option, timeboxed slack notifications

+++ backups/sources/source.py
@@ -1,6 +1,9 @@
+import os
+import time
+
+import backups.compress
 import backups.encrypt
 
-import os
 
 # Abstract
 class BackupSource:
@@ -15,14 +18,27 @@ class BackupSource:
             self.passphrase = config['passphrase']
         if 'tmpdir' in config:
             self.tmpdir = config['tmpdir']
+        self.compress = False
+        if 'compress_only' in config:
+            self.compress = config['compress_only'] == 1
 
-    def dump_and_compress(self):
+    def dump_and_compress(self, stats):
+        d_starttime = time.time()
         filenames = self.dump()
+        d_endtime = time.time()
+        stats.dumptime_dump = d_endtime - d_starttime
         if isinstance(filenames, basestring):
             filenames = [filenames, ]
-        encrypted_files = []
+        compressed_files = []
+
+        e_starttime = time.time()
         for filename in filenames:
-            encfilename = backups.encrypt.encrypt(filename, self.passphrase)
-            encrypted_files.append(encfilename)
-            os.unlink(filename)
-        return encrypted_files
+            if self.compress:
+                compressed_filename = backups.compress.compress(filename)
+            else:
+                compressed_filename = backups.encrypt.encrypt(filename, self.passphrase)
+                os.unlink(filename)
+            compressed_files.append(compressed_filename)
+        e_endtime = time.time()
+        stats.dumptime_encrypt = e_endtime - e_starttime
+        return compressed_files
